"Parser and Rules fixed"

+++ src/main/java/org/paterna/rules/rules/PflichtfeldRegel.java
@@ -5,9 +5,9 @@ public class PflichtfeldRegel extends Rule {
 	private String feldName1; // ID C3.2
 	private String wert1; // Steuerberichtigung einer Steuermeldung ab.
 	private String feldName2; // C65
-	private double wert2; // 0
+	private int wert2; // 0
 
-	public PflichtfeldRegel(String feldName1, String wert1, String feldName2, double wert2) {
+	public PflichtfeldRegel(String feldName1, String wert1, String feldName2, int wert2) {
 		this.feldName1 = feldName1;
 		this.wert1 = wert1;
 		this.feldName2 = feldName2;
@@ -16,7 +16,7 @@ public class PflichtfeldRegel extends Rule {
 
 	@Override
 	public String toXsdAssert() {
-		return String.format("if (%s = '%s' and %s &lt; %.2f) then true() else false()", feldName1, wert1, feldName2,
+		return String.format("if (%s = '%s' and %s &lt; %d) then true() else false()", feldName1, wert1, feldName2,
 				wert2);
 	}
 
@@ -33,7 +33,7 @@ public class PflichtfeldRegel extends Rule {
 		return feldName2;
 	}
 
-	public double getWert2() {
+	public int getWert2() {
 		return wert2;
 	}
 }
